refactor: move inline components to ui
